Assign security

+++ src/main/java/com/jelinski/niajee/motorcycle/service/MotorcycleService.java
@@ -4,11 +4,15 @@ import com.jelinski.niajee.motorcycle.entity.Motorcycle;
 import com.jelinski.niajee.motorcycle.repository.api.MotorcycleRepository;
 import com.jelinski.niajee.motorcycleType.entity.MotorcycleType;
 import com.jelinski.niajee.motorcycleType.repository.api.MotorcycleTypeRepository;
+import com.jelinski.niajee.user.entity.User;
+import com.jelinski.niajee.user.entity.UserRoles;
 import com.jelinski.niajee.user.repository.api.UserRepository;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ejb.EJBAccessException;
 import jakarta.ejb.LocalBean;
 import jakarta.ejb.Stateless;
 import jakarta.inject.Inject;
-import jakarta.transaction.Transactional;
+import jakarta.security.enterprise.SecurityContext;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
@@ -39,15 +43,27 @@ public class MotorcycleService {
     private final UserRepository userRepository;
 
     /**
+     * Security context
+     */
+    private final SecurityContext securityContext;
+
+    /**
      * @param motorcycleRepository     repository for motorcycle entity
      * @param motorcycleTypeRepository repository for motorcycleType entity
      * @param userRepository           repository for user entity
+     * @param securityContext          security context
      */
     @Inject
-    public MotorcycleService(MotorcycleRepository motorcycleRepository, MotorcycleTypeRepository motorcycleTypeRepository, UserRepository userRepository) {
+    public MotorcycleService(
+            MotorcycleRepository motorcycleRepository,
+            MotorcycleTypeRepository motorcycleTypeRepository,
+            UserRepository userRepository,
+            @SuppressWarnings("CdiInjectionPointsInspection") SecurityContext securityContext
+    ) {
         this.motorcycleRepository = motorcycleRepository;
         this.motorcycleTypeRepository = motorcycleTypeRepository;
         this.userRepository = userRepository;
+        this.securityContext = securityContext;
     }
 
     /**
@@ -56,27 +72,57 @@ public class MotorcycleService {
      * @param id motorcycle's id
      * @return container with motorcycle
      */
+    @RolesAllowed(UserRoles.USER)
     public Optional<Motorcycle> find(UUID id) {
-        return motorcycleRepository.find(id);
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return motorcycleRepository.find(id);
+        }
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+        return motorcycleRepository.findByIdAndUser(id, user);
+    }
+
+    /**
+     * @param id   character's id
+     * @param user existing user
+     * @return selected character for user
+     */
+    @RolesAllowed(UserRoles.USER)
+    public Optional<Motorcycle> find(User user, UUID id) {
+        return motorcycleRepository.findByIdAndUser(id, user);
     }
 
     /**
      * @return all available motorcycles
      */
+    @RolesAllowed(UserRoles.USER)
     public List<Motorcycle> findAll() {
-        return motorcycleRepository.findAll();
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return motorcycleRepository.findAll();
+        }
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+        return motorcycleRepository.findAllByUser(user);
     }
 
     /**
      * @return all available motorcycles by motorcycle type
      */
+    @RolesAllowed(UserRoles.ADMIN)
     public List<Motorcycle> findAll(MotorcycleType motorcycleType) {
         return motorcycleRepository.findAllByMotorcycleType(motorcycleType);
     }
 
+    @RolesAllowed(UserRoles.ADMIN)
     public Optional<List<Motorcycle>> findAllByMotorcycleType(UUID id) {
         return motorcycleTypeRepository.find(id)
                 .map(motorcycleRepository::findAllByMotorcycleType);
+    }
+
+    @RolesAllowed(UserRoles.ADMIN)
+    public Optional<List<Motorcycle>> findAllByUser(UUID id) {
+        return userRepository.find(id)
+                .map(motorcycleRepository::findAllByUser);
     }
 
     /**
@@ -84,6 +130,7 @@ public class MotorcycleService {
      *
      * @param motorcycle new motorcycle to be saved
      */
+    @RolesAllowed(UserRoles.ADMIN)
     public void create(Motorcycle motorcycle) {
         if (motorcycleRepository.find(motorcycle.getId()).isPresent()) {
             throw new IllegalArgumentException("Motorcycle already exists");
@@ -91,7 +138,26 @@ public class MotorcycleService {
         if (motorcycleTypeRepository.find(motorcycle.getMotorcycleType().getId()).isEmpty()) {
             throw new IllegalArgumentException("Motorcycle type does not exist");
         }
+
         motorcycleRepository.create(motorcycle);
+        motorcycleTypeRepository.find(motorcycle.getMotorcycleType().getId())
+                .ifPresent(motorcycleType -> motorcycleType.getMotorcycles().add(motorcycle));
+        userRepository.find(motorcycle.getUser().getId())
+                .ifPresent(user -> user.getMotorcycles().add(motorcycle));
+    }
+
+    /**
+     * Creates new motorcycle for current caller principal.
+     *
+     * @param motorcycle new motorcycle
+     */
+    @RolesAllowed(UserRoles.USER)
+    public void createForCallerPrincipal(Motorcycle motorcycle) {
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+
+        motorcycle.setUser(user);
+        create(motorcycle);
     }
 
     /**
@@ -99,7 +165,9 @@ public class MotorcycleService {
      *
      * @param motorcycle motorcycle to be updated
      */
+    @RolesAllowed(UserRoles.USER)
     public void update(Motorcycle motorcycle) {
+        checkAdminRoleOrOwner(motorcycleRepository.find(motorcycle.getId()));
         motorcycleRepository.update(motorcycle);
     }
 
@@ -108,8 +176,26 @@ public class MotorcycleService {
      *
      * @param id motorcycle's id
      */
+    @RolesAllowed(UserRoles.USER)
     public void delete(UUID id) {
+        checkAdminRoleOrOwner(motorcycleRepository.find(id));
         motorcycleRepository.delete(motorcycleRepository.find(id).orElseThrow());
+    }
+
+    /**
+     * @param motorcycle character to be checked
+     * @throws EJBAccessException when caller principal has no admin role and is not character's owner
+     */
+    private void checkAdminRoleOrOwner(Optional<Motorcycle> motorcycle) throws EJBAccessException {
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return;
+        }
+        if (securityContext.isCallerInRole(UserRoles.USER)
+                && motorcycle.isPresent()
+                && motorcycle.get().getUser().getLogin().equals(securityContext.getCallerPrincipal().getName())) {
+            return;
+        }
+        throw new EJBAccessException("Caller not authorized.");
     }
 
 }
